Move `ChainStatsCollator` to `chain_stats.rs`
