+++ backend/telemetry_core/src/state/mod.rs
@@ -15,6 +15,7 @@
 // along with this program. If not, see <https://www.gnu.org/licenses/>.
 
 mod chain;
+mod chain_stats;
 mod counter;
 mod node;
